修复：使用异步驱动时，已存储的值 == false 时不触发首次订阅

+++ kurimudb/data.ts
@@ -4,14 +4,12 @@ export default class Data {
       set: (target, key, value, proxy) => {
         key = model.checkPrimary(key);
         model.cache.put(key, value);
-        if (model.isPersistence())
-          (async () =>
-            await model.storage.insertOrUpdate(key, model.cache.get(key)))();
+        if (model.isPersistence()) (async () => await model.storage.insertOrUpdate(key, model.cache.get(key)))();
         model.changed.set(key);
         return true;
       },
       get: (target, key, receiver) => {
-        if ("string" === typeof key && key.endsWith("$")) {
+        if ('string' === typeof key && key.endsWith('$')) {
           // 取出原始值，无需 await
           key = key.substring(0, key.length - 1);
           key = model.checkPrimary(key);
@@ -25,7 +23,7 @@ export default class Data {
               // 值不存在，持久化，异步方式
               (async () => {
                 const value = await model.storage.select(key);
-                if (value) model.cache.put(key, value);
+                if (void 0 !== value) model.cache.put(key, value);
               })();
               const result = model.cache.subscribe(key);
               return result;
@@ -62,7 +60,7 @@ export default class Data {
         key = model.checkPrimary(key);
         if (model.options.async)
           throw new Error(
-            `For persistent models, the result of the "in" operator may be incorrect, need "await modelName.has(key)" replace your "key in modelName.data".`
+            `For persistent models, the result of the "in" operator may be incorrect, need "await modelName.has(key)" replace your "key in modelName.data".`,
           );
         return model.cache.has(key);
       },
@@ -114,8 +112,7 @@ export default class Data {
       deleteProperty: (target, key) => {
         key = model.checkPrimary(key);
         model.cache.forget(key);
-        if (model.isPersistence())
-          (async () => await model.storage.delete(key))();
+        if (model.isPersistence()) (async () => await model.storage.delete(key))();
         model.changed.set(key);
 
         return true;
